adding a matrix for saving known true pareto optimal solutions to the problem class and count of such solutions
a static function is added to the ustility class for calculating
euclidean distance between two real vectors

a variable is added for counting number of fitness evaluations and some
that was written for debugging the fast_nondominated_sort function is
omitted.

number of non-dominated solutions after termination of the algorithm in the
last population is calculated and printed to the console.

Signed-off-by: Shahab Abbasi Rad <darkoob12@gmail.com>